+++ features/page_objects/login_page.py
@@ -1,4 +1,6 @@
 LOGIN_BTN = '#login2'
+LOGIN_FORM_BTN = '#logInModal > div > div > div.modal-footer > button.btn.btn-primary'
+LOGOUT_BUTTON = '#logout2'
 
 LOGIN_URL = 'https://demoblaze.com/index.html'
 EXPECTEDHOME_URL = 'https://demoblaze.com/index.html'
