+++ features/utils/CommonConstants.py
@@ -1,4 +1,2 @@
 VALID_USERNAME = 'asdasdasd123asdasdasdasdasdasdasdasdasdasdasddfgdfhcserhgdfcgvdfg'
 VALID_PASSWORD = 'dsa'
-
-LOGOUT_BUTTON = '#logout2'
